pyproject.toml
[tool.poetry]
name = "sat-gpt-generator"
version = "0.1.0"
description = ""
authors = ["eldpswp99 <nrg1392@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.9"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
openai = "^0.28.1"
python-dotenv = "^1.0.0"
black = "^23.9.1"
langchain = "^0.0.306"
google-search-results = "^2.4.2"
numpy = "^1.26.0"
pandas = "^2.1.1"
openpyxl = "^3.1.2"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
